Updates to Responder Guard (By: David Fletcher)
David Fletcher: I refactored a bunch of code in the original ResponderGuard to make it a bit more modular. Then I implemented all of the changes that were suggested by Kevin (Inveigh). That included static NBNS source port and hostname control.
Then I created ResponderGuardAgent which can be used to send NBNS, LLMNR, or mDNS requests on a given subnet without sending unicast packets. The aim was to make it so you could have an agent deployed using GPO on targeted machines.
It works similar to ResponderGuard but the user can specify any combination of NBNS, LLMNR or mDNS request generation, can control the hostname, and can specify the honey creds during invocation.